requirements.txt
streamlit>=1.28.0
anthropic>=0.39.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0
reportlab>=4.0.0
pillow>=10.0.0
tabulate>=0.9.0
python-docx>=1.1.0
python-pptx>=0.6.21
matplotlib>=3.7.0
